refactor(audit): extract test modules from core_fingerprint + field_patterns to clear GodFiles
