+++ homeboy.json
@@ -176,7 +176,7 @@
           "src/core/code_audit/conventions/",
           "src/core/code_audit/comment_blocks.rs",
           "src/core/code_audit/comment_hygiene.rs",
-          "src/core/code_audit/core_fingerprint.rs",
+          "src/core/code_audit/core_fingerprint/mod.rs",
           "src/core/code_audit/core_fingerprint/relationships.rs",
           "src/core/code_audit/detectors/test_coverage.rs",
           "src/core/code_audit/detectors/global_env_guard.rs",
@@ -698,8 +698,6 @@
         "parallel-implementation::src/core/agent_task_prompts.rs::ParallelImplementation",
         "parallel-implementation::src/core/extension/trace/attach.rs::ParallelImplementation",
         "structural::src/commands::DirectorySprawl",
-        "structural::src/core/code_audit/core_fingerprint.rs::GodFile",
-        "structural::src/core/code_audit/detectors/field_patterns.rs::GodFile",
         "structural::src/core::DirectorySprawl",
         "test_quality::src/commands/bench/tests/scenario_run_tests.rs::VacuousTest",
         "test_quality::src/core/agent_task_scheduler/tests/outcome_tests.rs::VacuousTest",
